analyzer_app/requirements.txt
asgiref==3.4.1
attrs==21.2.0
click==8.0.1
fastapi==0.68.1
h11==0.12.0
iniconfig==1.1.1
Jinja2==3.0.1
loguru==0.5.3
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
